Code Cleanup: CvCivilizationInfo now mostly use enum arguments instead of ints
Cleaned up code calling those functions to get rid of a lot of typecasting.
As a bonus we also got rid of a bunch of getDefineINTs and replaced it with global const variables.
This means we will get a compile time error if any of those ml types are renamed or removed.
It's also faster, but that's just a bonus and properly called too rarely to have a real effect.