Fix: when an OOS occurs, the generate OOS log button will no longer toggle on/off
It used to only show if the OOS text was visible on the screen while opening the menu.
Now it shows if the text has been shown at least once the same turn.

As a nice bonus, we now have a function, which only reaches the end when an OOS occurs.
This will allow us to add code, which triggers automatically in case of an OOS.